fix(review): исправления замечаний code review крипто-системы
- CryptoInvoiceStatus enum вместо магических строк PENDING/PAID/EXPIRED
- Bulk UPDATE запрос для экспайра инвойсов вместо N+1
- int → long для spentThisMonth/avgPerMonth/spentAllTime (защита от переполнения)
- log.error с exception object (CryptoPaymentScheduler, CryptoPaymentBotListener)
- MAX_TOKENS=100_000 в TopupTextHandler — верхняя граница суммы пополнения
- Убрана мутация List внутри лямбды в TokenLedgerServiceImpl
- Единый parseMode="Markdown" в ControllerTokenStatusNotifier
- Collectors.joining и block(Duration) в CryptoBotClient

+++ valui-user/src/main/java/com/valui/user/dto/TokenInfoDto.java
@@ -8,7 +8,7 @@ public record TokenInfoDto(
     int monthlyTokenGrant,
     Integer tokenLowThreshold,
     List<TokenHistoryEntry> recentHistory,
-    int spentThisMonth,
-    int avgPerMonth,
-    int spentAllTime
+    long spentThisMonth,
+    long avgPerMonth,
+    long spentAllTime
 ) {}
